Switch to the new repo for test runner
This deletes all of the generalizable test runner code, which has been
copied to the new repo. However, the directory structure is left in
place, with modular-specific functionality remaining.

Note: lib/fidl/scope.h is not being deleted, but it has been copied to
the new repo. There will be a follow-up CL to find a single home for it.

Depends on:
https://fuchsia-review.googlesource.com/c/23845/
and:
https://fuchsia-review.googlesource.com/c/24016/

Change-Id: I1779bc594db58a8b55bd02de69951570cf440579