Simple storage for user shell.
Add a Link service to the UserShellContext. User shell can use this
to store state it wants to keep across invocations and across devices,
such as information about story clusters.

The Link is stored in the root page for now, under the key
"user-shell-link". It should be moved to its own page, and the
story info links should be moved to their own page too. Until
then, we prevent story provider impl to mistake the user shell
link for a story, as we did for the device map key.

We use the user shell link in dummy user shell.

Collateral cleanup:

* Add an option --test=0 to dummy_user_shell that lets it tolerate
  when it doesn't run under test runner. Otherwise it now always
  crashes when run directly, which is inconvenient for development.

Change-Id: Ie81525ffc1722c8868d3e9b9a7d2efa48dbfbeb1