Supply a "name" to a module when you start it.
- This "name" is used when naming/identifying links and message queues
on subsequent reinflations of a module. Particularly, the "module name
path" will be used to identify links and message queues; a module name
path is the path of module names that have started a module (ie.,
name[n] starts name[n+1]).
- A "null" name will name a module by its "component id" (ie. its URL)
- As a drive by, fix up some 'test.sh' paths to use the test runner
(they are currently pointed to a non-existent location).

When a story is reinflated, we need to identify the component instance
that a module was in its previous inflation so we can give it the same
message queue, link, etc. Right now, we naively assume identify this
using a module URL. Going forward, we would like to identify it using a
module name that the programmer assigns a module; this module's
link/message queue instances are identified using the 'path of this
module from the root', where each node in the path is the name a module
is assigned by the programmer.

Change-Id: I2f53248ee03227c2eab9c9acc49a91ee707f5b05