Move ledger ownership to UserRunnerImpl.
User runner creates the ledger, which is no longer
used exclusively by story provider. Therefore, user runner
retains ownership of the ledger connection and only hands
a refernece to the story provider. It is OK for the story
provider to run Operations against its ledger reference
in its own operation queue, because the life time of the
story provider (and hence its operation queue) is coupled
to the life time of the user runner.

Analog for the root page. It's opened by user runner, and
all its components, not just the story provider, write to
it, so user runner should own the connection.

With the change of ownership in the ledger to user runner,
ledger level operations, specifically registration of
a conflict resolver factory, can happen in the user runner,
which is the better place for it. The device map is
still written by the story provider; this changes soon.

Collateral cleanup:

* Remove public qualifier from base classes that aren't.

* const

* public

* Rearrange initialization order of fields.

Change-Id: I4380d9a6ca9358a3f2cf54f6befc1dfb4975a663