Migrate MessageQueueManager to Operation.
This ensures that MessageQueueManager can be deleted at any time,
while method invocations are pending on intermediate fidl services
used by deep-asynchronous methods, and those method invocations
are all properly cancelled.

This is a plain translation. Operations might get simplified in a
followup CL.

Asynchronous semantics is unchanged, we might want to switch it
to an OperationQueue in a followup.

Change-Id: I72a5018f223e2c41e623a5e2d8ce6208f621a9c4