Operations are inner classes.
That way they naturally have access to members of their host class.
This is not used yet, but can simplify initialization of an operation.

It also declares all operations in the class definition, which is
nice to read.

Also, gets rid of one last abuse of an anonymous namespace in a
header file.

Change-Id: I87af471ace0b87567007daa4de7be11a3b77e646