Ensure device_runner is a singleton.
The boot environment provides a service modular::DeviceRunnerMonitor.

When device_runner starts, it connects to this service and obtains
the number of connections from it. If that number is not one,
device_runner exits.

Just in case someone ever wants to run device_runner in a boot
environment that happens not to have such a service available,
the check can be bypassed with the device_runner command line
option --ignore_monitor.

FW-143 #done

Change-Id: I4d57cf640e56d7689302820662ada3df69149ba1