Cleanup while reading the code.
A method that adds a new connection to a service implementation
is sometimes called Connect(), sometimes AddBinding().

We settle for uniformly Connect(), because (1) it's the shorter
and simpler name, and (2) we often do more than simply adding a
binding, for example we also create a new ConnectionImpl where
we need to track the identity of the source of the connection
request, or we queue connection requests until the implementation
is fully initialized.

Change-Id: I77aef17ac2f755358014e3aad40324ffb0891218