StartModule takes link name instead of interface.
Module create/get links using ModuleContext.GetLink(<name>, ..), and
pass on links to child modules by passing on the name.

In the framework, we now use "link path" to reference a particular Link.
A link path consistents of a module path (of the module that "created"
the link), and the link name. A module can pass a link it has created to
a child module by passing on the link name. The framework will now
understand this relative name, translate it to the link path, and assign
that link path as

Additional changes underneath to accomodate this:
- All module (not just root modules) metadata is recorded in the ledger
  now: their module path, url, and their link name.
- Initialize() no longer gets a link interface, and must call
  ModuleContext.GetLink (we deprecate CreateLink).
- Chase down all uses of ModuleContext.StartModule() and
  Module.Initialize() and update them.
- Make StoryImpl::StartModule an operation (since it writes stuff to the
  ledger now).

Change-Id: Iedece8341666a2a68c8da79d8cc8b71e473c9d73