Switch bootstrap to ServiceNamspace
ServiceNamespace implements both ServiceProvider and svcfs, which means
after this change it will be possible to mount the bootstrap services in
the file system. However, this change doesn't actually use that ability.

After this fallback, bootstrap doesn't fall back to the root environment
for service lookup, but the root environment doesn't have any services,
so that isn't much of a loss. By removing the fallback, we gain the
ability to enumerate the bootstrap services, which might be useful
later.

Change-Id: Ie2c9fe9124ea10f9568281a08bb4506624ce3b5e