Better UserShell test
Rewrite test_user_shell as an asynchronous sequence of steps
verified by TestPoints. The sequence is almost linear now,
with the exception of a loop when we cycle a story between
running and stopped several times.

This explores the idea of tying one TestPoint to one succeeded
step, and writing one step as one method of a test class. The
test points are data members and written directly before the
test method they verify.

In order to properly tear down the test such that the test
point destructors can send their status to the test runner,
we change the test user shell implementation to delete itself
before sending TearDown() to the test runner.

In order to avoid the race condition between responding to
Terminate() to allow to being killed and sending further status
to the test runner, we delay the invocation of the Terminate()
response callback until after all test runner messages are
sent, which also after the tst user shell is deleted. Therefore,
we make it possible to extract the UserShell binding from
the implementation class and delete it later.

Collateral cleanup:

* Some typos.

* Comment clarifications.

* Simplify some code in the counter_cpp example.

Change-Id: I223675fbda08012d8d48b669db54b5aa2a5c9c0c