Better service termination.
Uniformly terminate services with a Terminate() method by waiting
for the service connection to close after invoking the Terminate()
method, and then disconnecting the application controller, rather
than to wait for the return callback of the Terminate() call.

This is done using a wrapper class AppClient which ties together the
parts of the termination sequence involving the app and the parts
involing a service.

As example, apply this to UserShell and maxwell services in user runner.
This also applies a timeout to the user shell Terminate() call.
Previously, if a user shell doesn't implement Terminate() properly,
UserRunner.Terminate() would hang.

More services to follow:
* DeviceShell
* UserRunner
* StoryShell
* Module
* Agent

With this in place, it will be very simple to drop the return callback
from the Terminate() method and continue at the close of the service
connection instead. This also uniformly applies a timeout.

Collateral cleanup:

* Clarify the termination sequence of the user runner implementation
  analog to what we did for test_user_shell: Ensure that the done
  callback of Terminate() is invoked after the instance is deleted.

Change-Id: Id47904f8fb9abb6040d5552cba394526b58bb356