FlowTokens for simpler Operations.
Implementing the idea by jimbe@ to automate invocation of
Done() by the destructor of a scoped instance:

At the start of Run(), we create an instance a token class
that holds a shared ref counter and a pointer to `this`
(i.e. the Operation instance).

The token is copyable, which increments the ref counter.

The destructor decrements the ref counter and calls Done()
when the counter reaches zero.

Now, if the token is placed onto the capture list of every
asynchronous method callback, then such a return callback
only needs to return in order trigger invocation of Done(),
or else if it passes the token on, Done() is deferred.

This also invokes Done() in the case that a method does not
invoke its callback because its connection gets closed,
which was not handled previously (and would cause the
operation queue to block).

Why this is ref counted, not moved:

1. Some flows of control branch into multiple parallel
   branches, and then its subtle to figure out which
   one to move it along.

2. To move something onto a capture list is more verbose
   than to just copy it, so it would defeat the purpose
   of being simpler to write than the status quo.

3. A lambda with something that is moveonly on its capture
   list is copyable anymore, and one of the points of
   Operation was to only have copyable continuations.

Example use in one Operation, more Operations will be changed
in followup.

Change-Id: I807a35d901c47e9afa6015bcacc28511bf55f5cf