Use incognito user by default in dev device shell.
After the recent oauth changes, if a user is not in the local
user db, device runner doesn't log them in, unless it's the
incognito user. And the default user1 is not the in the user db.

The modular tests all use the default user1, and therefore hang
without this fix.

Change-Id: I2b93d35efd091db7b54ac2d94cd3ce47cc43c32a